+++ src/main/java/com/pawandfeet/registration/exception/AddressNotFoundException.java
@@ -1,6 +1,6 @@
 package com.pawandfeet.registration.exception;
 
-public class AddressNotFoundException extends Exception {
+public class AddressNotFoundException extends RuntimeException {
 
     public AddressNotFoundException() {
         super("Address not found.");
